+++ monggregate/pipeline.py
@@ -392,7 +392,7 @@ class Pipeline(BaseModel): # pylint: disable=too-many-public-methods
 
         Arguments:
         ------------------------
-            - by / _id (offcial MongoDB name represented by a pydantic alias), str | list[str] | set[str] : field or group of fields to group on
+            - by,  str | list[str] | set[str] | dict | None : field or group of fields to group by
             - query, dict | None : Computed aggregated values (per group)
 
         """
@@ -668,7 +668,7 @@ class Pipeline(BaseModel): # pylint: disable=too-many-public-methods
         """
 
         self.stages.append(
-                ReplaceRoot(path=path)
+                ReplaceRoot(path=path, document=document)
             )
         return self
 
@@ -686,7 +686,7 @@ class Pipeline(BaseModel): # pylint: disable=too-many-public-methods
         """
 
         self.stages.append(
-                ReplaceRoot(path=path)
+                ReplaceRoot(path=path, document=document)
             )
         return self